Don't ignore the node_modules directory
As noted in the GitHub documentation:

GitHub downloads each action run in a workflow during runtime and
executes it as a complete package of code before you can use workflow
commands like run to interact with the runner machine. This means you
must include any package dependencies required to run the JavaScript
code. You'll need to check in the toolkit core and github packages to
your action's repository.